refactor: 추천글 도메인 import 변경 #131
