+++ src/main/java/com/bookripple/api/domain/recommendation/service/RecommendServiceImpl.java
@@ -1,15 +1,13 @@
 package com.bookripple.api.domain.recommendation.service;
 
-import com.bookripple.api.common.code.BookErrorCode;
-import com.bookripple.api.common.code.CommonErrorCode;
-import com.bookripple.api.common.code.RecommendErrorCode;
-import com.bookripple.api.common.error.ApiException;
+import com.bookripple.api.domain.book.code.BookErrorCode;
 import com.bookripple.api.domain.book.dto.BookRes;
 import com.bookripple.api.domain.book.entity.Book;
 import com.bookripple.api.domain.book.repository.BookRepository;
 import com.bookripple.api.domain.book.service.BookQueryService;
 import com.bookripple.api.domain.member.entity.Member;
 import com.bookripple.api.domain.member.repository.MemberRepository;
+import com.bookripple.api.domain.recommendation.code.RecommendErrorCode;
 import com.bookripple.api.domain.recommendation.converter.RecommendConverter;
 import com.bookripple.api.domain.recommendation.dto.RecommendReqDto.Create;
 import com.bookripple.api.domain.recommendation.dto.RecommendResDto.MyRecommend;
@@ -18,10 +16,12 @@ import com.bookripple.api.domain.recommendation.dto.RecommendResDto.Recommend;
 import com.bookripple.api.domain.recommendation.dto.RecommendResDto.RecommendList;
 import com.bookripple.api.domain.recommendation.entity.Recommendation;
 import com.bookripple.api.domain.recommendation.repository.RecommendRepository;
+import com.bookripple.api.global.code.CommonErrorCode;
 import com.bookripple.api.global.converter.GlobalConverter;
 import com.bookripple.api.global.dto.GlobalDto.ContentReq;
 import com.bookripple.api.global.dto.GlobalDto.IdList;
 import com.bookripple.api.global.dto.GlobalDto.IdRes;
+import com.bookripple.api.global.error.ApiException;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
